sad paths delete review

+++ models/reviews.js
@@ -26,6 +26,11 @@ exports.insertReview = async (guest_id, rating, comment, id) => {
 }
 
 exports.deleteReview = async(id) => {
+    const { rows: [checkExist] } = await db.query(`SELECT * FROM reviews WHERE review_id = $1`, [id])
+    console.log(checkExist)
+    if(checkExist === undefined){
+        return checkExist
+    }
     const { rows: review } = await db.query(`DELETE FROM reviews 
         WHERE review_id = $1`, [id]
     )
